Implement SplashScreen and restructure navigation layout; remove old default tab layout. Added Copy-All for debug
