+++ app/DebugScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ScrollView, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { ScrollView, Text, TouchableOpacity, StyleSheet, Alert, Clipboard } from 'react-native';
 import { useDebug } from '../src/DebugContext';
 import { useRouter } from 'expo-router';
 
@@ -7,10 +7,19 @@ const DebugScreen: React.FC = () => {
   const { messages } = useDebug();
   const router = useRouter();
 
+  const copyToClipboard = () => {
+    const allMessages = messages.join('\n');
+    Clipboard.setString(allMessages);
+    Alert.alert('Copied to Clipboard', 'All debug messages have been copied to the clipboard.');
+  };
+
   return (
     <ScrollView contentContainerStyle={styles.container}>
       <TouchableOpacity onPress={() => router.back()} style={styles.backButton}>
         <Text style={styles.backButtonText}>Back</Text>
+      </TouchableOpacity>
+      <TouchableOpacity onPress={copyToClipboard} style={styles.copyButton}>
+        <Text style={styles.copyButtonText}>Copy All</Text>
       </TouchableOpacity>
       {messages.map((message, index) => (
         <Text key={index} selectable style={styles.messageText}>
@@ -26,15 +35,25 @@ const styles = StyleSheet.create({
     padding: 10,
   },
   backButton: {
-    marginBottom: 20,
+    marginBottom: 10,
+    padding: 10,
+    backgroundColor: '#ccc',
+    borderRadius: 5,
   },
   backButtonText: {
-    color: 'blue',
-    fontSize: 18,
+    color: '#000',
+  },
+  copyButton: {
+    marginBottom: 10,
+    padding: 10,
+    backgroundColor: '#007bff',
+    borderRadius: 5,
+  },
+  copyButtonText: {
+    color: '#fff',
   },
   messageText: {
-    marginBottom: 10,
-    fontSize: 16,
+    marginBottom: 5,
   },
 });
